src/requirements.txt
tradingview-ta==3.3.0
pydantic-settings==2.8.1
tinkoff-investments==0.2.0b110
pandas==2.2.3
python-telegram-bot